Refactored and renamed getAbility and getPokemon to clarify JSON grabbing
